# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: StudentHub Africa API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List services
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceList"

    post:
      operationId: createService
      tags: [services]
      summary: Create a service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
...
  /services/{id}:
    get:
      operationId: getService
      tags: [services]
      summary: Get a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "404":
          description: Not found

    patch:
      operationId: updateService
      tags: [services]
      summary: Update a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"

    delete:
      operationId: deleteService
      tags: [services]
      summary: Delete a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # MISSIONS
  /missions:
    get:
      operationId: listMissions
      tags: [missions]
      summary: List missions
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, in_progress, completed, cancelled]
        - name: skill
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: minBudget
          in: query
          schema:
            type: number
        - name: maxBudget
          in: query
          schema:
            type: number
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MissionList"

    post:
      operationId: createMission
      tags: [missions]
      summary: Create a mission
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MissionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mission"
...
  /missions/{id}:
    get:
      operationId: getMission
      tags: [missions]
      summary: Get a mission
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mission"
        "404":
          description: Not found

    patch:
      operationId: updateMission
      tags: [missions]
      summary: Update a mission
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MissionUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mission"

    delete:
      operationId: deleteMission
      tags: [missions]
      summary: Delete a mission
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # APPLICATIONS
...
  /reviews:
    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /reviews/user/{userId}:
    get:
      operationId: listUserReviews
      tags: [reviews]
      summary: List reviews for a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"

  # DASHBOARD
...
components:
  schemas:
...
    User:
      type: object
      required: [id, email, name, role, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [student, company, individual]
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        university:
          type: ["string", "null"]
        skills:
          type: array
          items:
            type: string
        averageRating:
          type: ["number", "null"]
        totalReviews:
          type: integer
        createdAt:
          type: string
          format: date-time
...
    Service:
      type: object
      required: [id, title, description, price, deliveryDays, category, studentId, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        price:
          type: number
        deliveryDays:
          type: integer
        category:
          type: string
        imageUrl:
          type: ["string", "null"]
        totalOrders:
          type: integer
        averageRating:
          type: ["number", "null"]
        studentId:
          type: integer
        student:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
          format: date-time
    ServiceList:
      type: object
      required: [services, total]
      properties:
        services:
          type: array
          items:
            $ref: "#/components/schemas/Service"
        total:
          type: integer
    ServiceInput:
      type: object
      required: [title, description, price, deliveryDays, category]
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        deliveryDays:
          type: integer
        category:
          type: string
        imageUrl:
          type: string
    ServiceUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        deliveryDays:
          type: integer
        category:
          type: string
        imageUrl:
          type: string

    # MISSION
    Mission:
      type: object
      required: [id, title, description, budget, deadline, status, clientId, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        budget:
          type: number
        deadline:
          type: string
          format: date
        status:
          type: string
          enum: [open, in_progress, completed, cancelled]
        skills:
          type: array
          items:
            type: string
        clientId:
          type: integer
        client:
          $ref: "#/components/schemas/User"
        applicationCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    MissionList:
      type: object
      required: [missions, total]
      properties:
        missions:
          type: array
          items:
            $ref: "#/components/schemas/Mission"
        total:
          type: integer
    MissionInput:
      type: object
      required: [title, description, budget, deadline]
      properties:
        title:
          type: string
        description:
          type: string
        budget:
          type: number
        deadline:
          type: string
          format: date
        skills:
          type: array
          items:
            type: string
    MissionUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        budget:
          type: number
        deadline:
          type: string
          format: date
        status:
          type: string
          enum: [open, in_progress, completed, cancelled]
        skills:
          type: array
          items:
            type: string

    # APPLICATION
...
    Review:
      type: object
      required: [id, reviewerId, revieweeId, rating, createdAt]
      properties:
        id:
          type: integer
        reviewerId:
          type: integer
        reviewer:
          $ref: "#/components/schemas/User"
        revieweeId:
          type: integer
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    ReviewInput:
      type: object
      required: [revieweeId, rating]
      properties:
        revieweeId:
          type: integer
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string

    # DASHBOARD
...