# lib/api-zod/src/generated/generated/types/missionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * StudentHub Africa API
 * OpenAPI spec version: 0.1.0
 */

export interface MissionInput {
  title: string;
  description: string;
  budget: number;
  deadline: Date;
  skills?: string[];
}