# lib/api-zod/src/generated/types/serviceUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * StudentHub Africa API
 * OpenAPI spec version: 0.1.0
 */

export interface ServiceUpdate {
  title?: string;
  description?: string;
  price?: number;
  deliveryDays?: number;
  category?: string;
  imageUrl?: string;
}